Time: 347 ms (65.14%), Space: 13.9 MB (88.47%) - LeetHub

+++ 718-maximum-length-of-repeated-subarray/718-maximum-length-of-repeated-subarray.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     int n, m;
-    int dp[1001][1001];
+    short dp[1001][1001];
     int findLength(vector<int>& nums1, vector<int>& nums2) {
         n = nums1.size(), m = nums2.size();
         
-        int ans = 0;
+        short ans = 0;
         for(int i = 1; i <= n; i++){
             for(int j = 1; j <= m; j++) {
                 dp[i][j] = (nums1[i-1] == nums2[j-1] ? 1 + dp[i-1][j-1] : 0);
